+++ packages/pwa/src/routes/party_.$partyId.pay.tsx
@@ -86,7 +86,7 @@ function RouteComponent() {
     <div className="flex min-h-full flex-col">
       <div className="container flex h-16 items-center px-2 mt-safe">
         <BackButton fallbackOptions={{ to: "/party/$partyId" }} />
-        <h1 className="pl-4 text-2xl font-bold">
+        <h1 className="max-h-12 truncate px-4 text-xl font-medium">
           {isFromMe ? <Trans>Pay</Trans> : <Trans>Mark as paid</Trans>}
         </h1>
       </div>