+++ packages/pwa/src/routes/party_.$partyId.who.tsx
@@ -86,7 +86,7 @@ function Who() {
       <div className="container flex h-16 items-center px-2 mt-safe">
         <BackButton fallbackOptions={{ to: "/party/$partyId" }} />
 
-        <h1 className="pl-4 text-2xl font-bold">
+        <h1 className="max-h-12 truncate px-4 text-xl font-medium">
           <Trans>Who are you?</Trans>
         </h1>