+++ packages/pwa/src/routes/privacy-policy.tsx
@@ -11,7 +11,7 @@ function PrivacyPolicy() {
       <div className="container flex h-16 items-center px-2 mt-safe">
         <BackButton fallbackOptions={{ to: "/about" }} />
 
-        <h1 className="pl-4 text-2xl font-bold text-accent-900 dark:text-accent-100">
+        <h1 className="max-h-12 truncate px-4 text-xl font-medium text-accent-900 dark:text-accent-100">
           Privacy Policy
         </h1>
       </div>